requirements.txt
annotated-types==0.7.0
anyio==4.4.0
Authlib==1.3.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
exceptiongroup==1.2.1
grpcio==1.64.1
grpcio-health-checking==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
protobuf==5.27.2
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.2
validators==0.28.3
weaviate-client==4.6.5
